+++ apps/handcraft/src/petstore.rs
